Füge Layer.h, Layer.cpp, Neuron.h und Neuron.cpp hinzu
