+++ .gitignore
@@ -67,3 +67,10 @@ build/CMakeFiles/NN.dir/src/main.cpp.o
 build/CMakeFiles/NN.dir/src/main.cpp.o.d
 build/CMakeFiles/NN.dir/src/NeuralNetwork.cpp.o
 build/CMakeFiles/NN.dir/src/NeuralNetwork.cpp.o.d
+build/.cmake/api/v1/reply/codemodel-v2-d2cf5fd610c49f7054d5.json
+build/.cmake/api/v1/reply/index-2024-04-23T14-49-26-0341.json
+build/.cmake/api/v1/reply/target-NN-Debug-920bc10ada02d8542d05.json
+build/CMakeFiles/NN.dir/src/Layer.cpp.o
+build/CMakeFiles/NN.dir/src/Layer.cpp.o.d
+build/CMakeFiles/NN.dir/src/Neuron.cpp.o
+build/CMakeFiles/NN.dir/src/Neuron.cpp.o.d
